PDNC updated to use next-point
This is the final update that makes PDNC use 'current and next' points for drawing the curve, leading to a connected polygon-line instead of the dissconnected line it previously drew. That means that even with very short days (1800 ticks) and very slow PDNC tickrate (240 ticks) it will still draw the correct, connected lines!